requirements.txt
chromadb>=0.4.0
langchain>=0.0.300
sentence-transformers>=2.2.2
beautifulsoup4>=4.12.0
requests>=2.31.0
openai>=1.0.0
pypdf2>=3.0.0
tiktoken>=0.4.0
python-dotenv>=1.0.0
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
pymupdf>=1.23.0
streamlit-chat>=0.0.2
pillow>=11.3.0
